+++ bootstrap-core/src/main/java/de/agilecoders/wicket/core/markup/html/bootstrap/behavior/AssertTagNameBehavior.java
@@ -1,7 +1,7 @@
 package de.agilecoders.wicket.core.markup.html.bootstrap.behavior;
 
 import de.agilecoders.wicket.core.util.Components;
-import de.agilecoders.wicket.jquery.Generics2;
+import de.agilecoders.wicket.jquery.util.Generics2;
 import org.apache.wicket.Component;
 import org.apache.wicket.behavior.Behavior;
 import org.apache.wicket.markup.ComponentTag;
