+++ bootstrap-core/src/main/java/de/agilecoders/wicket/core/markup/html/bootstrap/block/prettyprint/PrettyJavaScriptReference.java
@@ -1,7 +1,7 @@
 package de.agilecoders.wicket.core.markup.html.bootstrap.block.prettyprint;
 
 import de.agilecoders.wicket.core.Bootstrap;
-import de.agilecoders.wicket.jquery.Generics2;
+import de.agilecoders.wicket.jquery.util.Generics2;
 import org.apache.wicket.markup.head.CssHeaderItem;
 import org.apache.wicket.markup.head.HeaderItem;
 import org.apache.wicket.markup.head.JavaScriptHeaderItem;
