+++ bootstrap-core/src/main/java/de/agilecoders/wicket/core/util/CssClassNames.java
@@ -3,8 +3,8 @@ package de.agilecoders.wicket.core.util;
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import de.agilecoders.wicket.core.markup.html.bootstrap.behavior.ICssClassNameProvider;
-import de.agilecoders.wicket.jquery.Generics2;
-import de.agilecoders.wicket.jquery.Strings2;
+import de.agilecoders.wicket.jquery.util.Generics2;
+import de.agilecoders.wicket.jquery.util.Strings2;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.util.lang.Args;
