+++ bootstrap-core/src/test/java/de/agilecoders/wicket/core/markup/html/bootstrap/behavior/CssClassNameModifierTest.java
@@ -2,7 +2,7 @@ package de.agilecoders.wicket.core.markup.html.bootstrap.behavior;
 
 import de.agilecoders.wicket.core.WicketApplicationTest;
 import de.agilecoders.wicket.core.test.IntegrationTest;
-import de.agilecoders.wicket.jquery.Generics2;
+import de.agilecoders.wicket.jquery.util.Generics2;
 import org.apache.wicket.AttributeModifier;
 import org.apache.wicket.Component;
 import org.apache.wicket.markup.html.WebMarkupContainer;
