+++ bootstrap-core/src/test/java/de/agilecoders/wicket/core/util/CssClassNamesTest.java
@@ -1,7 +1,7 @@
 package de.agilecoders.wicket.core.util;
 
 import de.agilecoders.wicket.core.test.TestCategory;
-import de.agilecoders.wicket.jquery.Generics2;
+import de.agilecoders.wicket.jquery.util.Generics2;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
