+++ bootstrap-themes/src/main/java/de/agilecoders/wicket/themes/markup/html/google/GoogleCssReference.java
@@ -1,7 +1,7 @@
 package de.agilecoders.wicket.themes.markup.html.google;
 
 import de.agilecoders.wicket.core.markup.html.themes.bootstrap.BootstrapCssReference;
-import de.agilecoders.wicket.jquery.Generics2;
+import de.agilecoders.wicket.jquery.util.Generics2;
 
 import org.apache.wicket.markup.head.CssHeaderItem;
 import org.apache.wicket.markup.head.HeaderItem;
